@@ -1,9 +1,9 @@
 {
     "require": {
-        "symfony/dom-crawler": "^6.4",
-        "symfony/css-selector": "^6.4",
-        "symfony/serializer": "^6.4",
-        "symfony/property-access": "^6.4",
-        "symfony/property-info": "^6.4"
+        "symfony/dom-crawler": "^7.1",
+        "symfony/css-selector": "^7.1",
+        "symfony/serializer": "^7.1",
+        "symfony/property-access": "^7.1",
+        "symfony/property-info": "^7.1"
     }
 }
