@@ -16,13 +16,13 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "symfony/polyfill-ctype": "~1.8",
         "symfony/polyfill-mbstring": "~1.0",
         "masterminds/html5": "^2.6"
     },
     "require-dev": {
-        "symfony/css-selector": "^5.4|^6.0|^7.0"
+        "symfony/css-selector": "^6.4|^7.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\DomCrawler\\": "" },
