@@ -16,12 +16,11 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
-        "symfony/deprecation-contracts": "^2.5|^3",
-        "symfony/property-info": "^5.4|^6.0|^7.0"
+        "php": ">=8.2",
+        "symfony/property-info": "^6.4|^7.0"
     },
     "require-dev": {
-        "symfony/cache": "^5.4|^6.0|^7.0"
+        "symfony/cache": "^6.4|^7.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\PropertyAccess\\": "" },
