@@ -16,44 +16,44 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/polyfill-ctype": "~1.8"
     },
     "require-dev": {
-        "doctrine/annotations": "^1.12|^2",
-        "seld/jsonlint": "^1.10",
         "phpdocumentor/reflection-docblock": "^3.2|^4.0|^5.0",
-        "symfony/cache": "^5.4|^6.0|^7.0",
-        "symfony/config": "^5.4|^6.0|^7.0",
-        "symfony/console": "^5.4|^6.0|^7.0",
-        "symfony/dependency-injection": "^5.4|^6.0|^7.0",
-        "symfony/error-handler": "^5.4|^6.0|^7.0",
-        "symfony/filesystem": "^5.4|^6.0|^7.0",
-        "symfony/form": "^5.4|^6.0|^7.0",
-        "symfony/http-foundation": "^5.4|^6.0|^7.0",
-        "symfony/http-kernel": "^5.4|^6.0|^7.0",
-        "symfony/messenger": "^5.4|^6.0|^7.0",
-        "symfony/mime": "^5.4|^6.0|^7.0",
-        "symfony/property-access": "^5.4.26|^6.3|^7.0",
-        "symfony/property-info": "^5.4.24|^6.2.11|^7.0",
+        "phpstan/phpdoc-parser": "^1.0|^2.0",
+        "seld/jsonlint": "^1.10",
+        "symfony/cache": "^6.4|^7.0",
+        "symfony/config": "^6.4|^7.0",
+        "symfony/console": "^6.4|^7.0",
+        "symfony/dependency-injection": "^7.2",
+        "symfony/error-handler": "^6.4|^7.0",
+        "symfony/filesystem": "^6.4|^7.0",
+        "symfony/form": "^6.4|^7.0",
+        "symfony/http-foundation": "^6.4|^7.0",
+        "symfony/http-kernel": "^6.4|^7.0",
+        "symfony/messenger": "^6.4|^7.0",
+        "symfony/mime": "^6.4|^7.0",
+        "symfony/property-access": "^6.4|^7.0",
+        "symfony/property-info": "^6.4|^7.0",
         "symfony/translation-contracts": "^2.5|^3",
-        "symfony/uid": "^5.4|^6.0|^7.0",
+        "symfony/type-info": "^7.1",
+        "symfony/uid": "^6.4|^7.0",
         "symfony/validator": "^6.4|^7.0",
-        "symfony/var-dumper": "^5.4|^6.0|^7.0",
-        "symfony/var-exporter": "^5.4|^6.0|^7.0",
-        "symfony/yaml": "^5.4|^6.0|^7.0"
+        "symfony/var-dumper": "^6.4|^7.0",
+        "symfony/var-exporter": "^6.4|^7.0",
+        "symfony/yaml": "^6.4|^7.0"
     },
     "conflict": {
-        "doctrine/annotations": "<1.12",
         "phpdocumentor/reflection-docblock": "<3.2.2",
         "phpdocumentor/type-resolver": "<1.4.0",
-        "symfony/dependency-injection": "<5.4",
-        "symfony/property-access": "<5.4",
-        "symfony/property-info": "<5.4.24|>=6,<6.2.11",
+        "symfony/dependency-injection": "<6.4",
+        "symfony/property-access": "<6.4",
+        "symfony/property-info": "<6.4",
+        "symfony/uid": "<6.4",
         "symfony/validator": "<6.4",
-        "symfony/uid": "<5.4",
-        "symfony/yaml": "<5.4"
+        "symfony/yaml": "<6.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Serializer\\": "" },
