@@ -16,18 +16,19 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "symfony/polyfill-ctype": "~1.8",
         "symfony/polyfill-intl-grapheme": "~1.0",
         "symfony/polyfill-intl-normalizer": "~1.0",
         "symfony/polyfill-mbstring": "~1.0"
     },
     "require-dev": {
-        "symfony/error-handler": "^5.4|^6.0|^7.0",
-        "symfony/intl": "^6.2|^7.0",
-        "symfony/http-client": "^5.4|^6.0|^7.0",
+        "symfony/error-handler": "^6.4|^7.0",
+        "symfony/emoji": "^7.1",
+        "symfony/http-client": "^6.4|^7.0",
+        "symfony/intl": "^6.4|^7.0",
         "symfony/translation-contracts": "^2.5|^3.0",
-        "symfony/var-exporter": "^5.4|^6.0|^7.0"
+        "symfony/var-exporter": "^6.4|^7.0"
     },
     "conflict": {
         "symfony/translation-contracts": "<2.5"
