@@ -15,18 +15,17 @@
     "GPL-2.0+"
   ],
   "require": {
-    "typo3/cms-core": "^12.4",
-    "symfony/dom-crawler": "^6.4 || ^7.0",
-    "symfony/css-selector": "^6.4 || ^7.0",
-    "symfony/serializer": "^6.4 || ^7.0",
-    "symfony/property-access": "^6.4 || ^7.0",
-    "symfony/property-info": "^6.4 || ^7.0",
+    "typo3/cms-core": "^13.4",
+    "symfony/dom-crawler": "^7.1",
+    "symfony/css-selector": "^7.1",
+    "symfony/serializer": "^7.1",
+    "symfony/property-access": "^7.1",
+    "symfony/property-info": "^7.1",
     "ext-dom": "*"
   },
   "require-dev": {
-    "typo3/testing-framework": "^8.0.0",
-    "friendsofphp/php-cs-fixer": "^3.0",
-    "ssch/typo3-rector": "^3.1"
+    "typo3/testing-framework": "^9.0",
+    "friendsofphp/php-cs-fixer": "^3.0"
   },
   "replace": {
     "typo3-ter/uw-a11y-check": "self.version"
